% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicktools-package.R
\docType{package}
\name{nicktools-package}
\alias{nicktools}
\alias{nicktools-package}
\title{nicktools: Nick's First R Package}
\description{
A suite of toy functions and data sets
}
\author{
\strong{Maintainer}: Nick Rasmussen \email{nicholas.rasmussen@water.ca.gov} (\href{https://orcid.org/0000-0002-5106-9038}{ORCID})

}
\keyword{internal}
